+++ CHANGELOG.md
@@ -6,11 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+### Added
+
+- `SNARIMAX` to `time_series`. This is a generic model which encompasses time series models such as ARIMA and NARX.
+
 ## [0.4.3](https://pypi.org/project/creme/0.4.3/) - 2019-10-27
 
 ### Added
 
 - `fetch_credit_card` to `datasets`, which streams a highly imbalanced dataset of fraudulent credit card transactions
+- The `utils` module now has a `math` submodule
 
 ### Fixed